+++ NEWS.md
@@ -1,3 +1,9 @@
+# fanyi 0.0.8
+
++ update citation information (2024-08-21, Wed)
++ fixed issue of 'chatglm' (2024-08-07, Wed)
++ update `ydict()` as youdao change its content. (2024-05-30, Thu)
+
 # fanyi 0.0.7
 
 + add claims in the DESCRIPTION that this package is not affiliated with any of the online translators and we take no responsibility (2024-04-14, Sun)